backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
black==26.3.1
brotli==1.2.0
cbor2==5.8.0
celery==5.6.2
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==46.0.5
cssselect2==0.9.0
daphne==4.2.1
dj-database-url==3.1.2
Django==6.0.3
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-jazzmin==3.0.3
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
flower==2.0.1
fonttools==4.62.0
gunicorn==25.1.0
humanize==4.15.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
msgpack==1.1.2
mypy_extensions==1.1.0
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.4
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydyf==0.12.1
PyJWT==2.11.0
pyOpenSSL==25.3.0
pyphen==0.17.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytokens==0.4.1
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.3.0
referencing==0.37.0
rpds-py==0.30.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.1.0
tornado==6.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.12.0
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.6.0
weasyprint==68.1
webencodings==0.5.1
whitenoise==6.12.0
zope.interface==8.2
zopfli==0.4.1
